# api/openapi.yaml
openapi: 3.1.0
info:
  title: OCS Testbench API
  version: 0.2.0
  description: |
    REST API for the OCS Testbench UI and external clients.

    This specification is the single source of truth for the HTTP contract
    between the Go backend (`internal/api`) and the React frontend (`web/`).
    Generated TypeScript types live at `web/src/api/schema.d.ts` — regenerate
    after any change to this file.

    See `docs/ARCHITECTURE.md` for behavioural semantics (state machines,
    invariants, variable resolution rules). This document is the wire
    contract; it does not restate the semantics.

    Real-time server → client events (peer status changes, execution
    progress, pause / resume transitions) are delivered via the SSE
    endpoint at `GET /events`.
servers:
  - url: /api/v1
    description: Current server (relative to origin)
...
paths:
...
  /executions:
    get:
      tags: [Executions]
      operationId: listExecutions
      summary: List past and in-progress executions
      parameters:
        - name: state
          in: query
          description: Filter by execution state
          required: false
          schema:
            $ref: '#/components/schemas/ExecutionState'
        - name: scenarioId
          in: query
          required: false
          schema:
            type: string
        - name: batchId
          in: query
          description: Group executions started together by a batched Run
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Execution list (newest first)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Executions]
      operationId: startExecution
      summary: Start a scenario — single run or batched
      description: |
        Creates one or more executions of a scenario. With defaults
        (`concurrency: 1`, `repeats: 1`) a single execution is returned in
        `items`. With higher values the engine spawns `concurrency`
        parallel workers, each running the scenario `repeats` times; all
        executions share a common `batchId` and are returned in `items`.

        `mode` controls runtime behaviour:
        - `interactive` — honours `pause` steps and runtime breakpoints.
          Typical when a user starts from the Scenario Builder.
        - `continuous` — ignores pauses and breakpoints; runs straight
          through. Required for batched runs.

        Optional `overrides` apply to every spawned execution:
        - `subscriberIds` — pool, round-robin across iterations
        - `peerId` — overrides the scenario's default peer binding for
          this run only

        `concurrency` and `repeats` only apply to `continuous` mode and
        are ignored (rejected with 422) for `interactive` runs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExecutionInput'
      responses:
        '202':
          description: Execution started; returns the session identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartExecutionSimpleResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/Problem'
  /executions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Executions]
      operationId: getExecution
      summary: Get a single execution with its full step-by-step detail
      description: |
        Returns the current status of an execution session, including
        lifecycle state, current step index, and aggregated timing
        metrics. For running executions this is a point-in-time snapshot;
        real-time updates are delivered via the SSE
        `GET /events/executions/{id}` endpoint.
      responses:
        '200':
          description: Execution status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStatusResult'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
...
  /executions/{id}/step:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Executions]
      operationId: stepExecution
      summary: Advance exactly one step, then pause again
      description: |
        Only valid from `paused` state. Executes exactly one scenario step
        and returns to `paused`. Returns the step result including result
        code and assertion outcomes.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepExecutionInput'
      responses:
        '200':
          description: Step executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepExecutionResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IllegalStateProblem'
        default:
          $ref: '#/components/responses/Problem'
  /executions/{id}/stop:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Executions]
      operationId: stopExecution
      summary: Stop a running execution after the current step
      description: |
        Requests the engine to stop the execution cleanly. The execution
        finishes its current step and then transitions to `completed` or
        `terminated`. Returns immediately — the actual stop happens
        asynchronously. This is the clean-stop variant; use `/abort`
        for immediate forceful termination.
      responses:
        '200':
          description: Stop accepted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
...
  /executions/{id}/interrupt:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Executions]
      operationId: interruptExecution
      summary: Interrupt a continuous execution after the current send
      description: |
        Signals the engine to pause after the current iteration completes.
        Only valid for continuous-mode executions in `active` state.
        The execution transitions to `paused`; use `/run-to-end` to resume.
      responses:
        '200':
          description: Interrupt accepted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IllegalStateProblem'
        default:
          $ref: '#/components/responses/Problem'
  /executions/{id}/run-to-end:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Executions]
      operationId: runToEnd
      summary: Resume an interrupted execution to completion
      description: |
        Resumes a continuous execution that was interrupted via `/interrupt`,
        running the remaining steps to completion. Also triggers the next
        step for interactive-mode executions in `paused` state.
      responses:
        '200':
          description: Resumed
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IllegalStateProblem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      description: Maximum number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    OffsetQuery:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Problem:
      description: Error response (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationProblem:
      description: |
        Validation error (RFC 7807). Carries field-level error messages
        keyed by JSON Pointer references into the request body.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IllegalStateProblem:
      description: |
        The requested operation is not legal from the execution's current
        state. For example: resuming an execution that is not `paused`,
        or aborting one that has already reached a terminal state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    PageMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0

    # ─────────────── Dashboard ───────────────
...
    ExecutionMode:
      type: string
      description: |
        Runtime behaviour of an execution — independent of the scenario's
        `sessionMode`.

        - `interactive` — honours pause steps and runtime breakpoints;
                           UI presents a debugger.
        - `continuous`  — ignores pauses and breakpoints; runs through.
                           Required for batched runs (`concurrency > 1`
                           or `repeats > 1`).
      enum: [interactive, continuous]
    ExecutionState:
      type: string
      description: |
        Execution state machine. See docs/ARCHITECTURE.md §6.

        Non-terminal: `pending`, `running`, `paused`.
        Terminal: `success`, `failure`, `aborted`, `error`.
      enum:
        - pending
        - running
        - paused
        - success
        - failure
        - aborted
        - error
...
    StartExecutionInput:
      type: object
      required: [scenarioId, mode]
      properties:
        scenarioId:
          type: string
        mode:
          $ref: '#/components/schemas/ExecutionMode'
        overrides:
          type: object
          description: Per-run overrides of the scenario's default bindings.
          properties:
            subscriberIds:
              type: array
              description: |
                Subscriber pool. When multiple ids are supplied and
                `repeats > 1`, the engine round-robins across iterations.
              items:
                type: string
              minItems: 1
            peerId:
              type: string
              description: |
                Overrides the scenario's bound peer for this run only.
        concurrency:
          type: integer
          description: |
            Parallel workers. Only legal with `mode: continuous`; values
            other than 1 with `interactive` return 422.
          minimum: 1
          default: 1
        repeats:
          type: integer
          description: |
            Iterations per worker. Only legal with `mode: continuous`.
          minimum: 1
          default: 1
    StartExecutionSimpleResult:
      type: object
      required: [sessionId]
      description: |
        Simplified response for `POST /executions`. Returns the session
        ID that can be used with `/executions/{id}` and
        `/events/executions/{id}`.
      properties:
        sessionId:
          type: string
          description: Opaque identifier for the started execution session.
    ExecutionStatusResult:
      type: object
      required: [sessionId, state, currentStep, metrics]
      description: |
        Current status of an execution session returned by
        `GET /executions/{id}`.
      properties:
        sessionId:
          type: string
        state:
          type: string
          description: |
            Lifecycle state: `active`, `paused`, `completed`,
            `terminated`, or `error`.
        currentStep:
          type: integer
          minimum: 0
          description: 0-based index of the current or last-completed step.
        metrics:
          $ref: '#/components/schemas/ExecutionMetricsResult'
...
    StepExecutionInput:
      type: object
      description: |
        Optional body for `POST /executions/{id}/step`. Supplies
        per-step variable overrides that shadow the session context for
        this one step only.
      properties:
        overrides:
          type: object
          additionalProperties: true
          description: Variable name → value map for this step only.
    StepExecutionResult:
      type: object
      required: [stepIndex, skipped, resultCode, assertionsPassed, assertions]
      description: |
        Result of a single step execution returned by
        `POST /executions/{id}/step`.
      properties:
        stepIndex:
          type: integer
          minimum: 0
          description: 0-based index of the step that executed.
        skipped:
          type: boolean
          description: True when a guard prevented the CCR from being sent.
        resultCode:
          type: integer
          minimum: 0
          description: |
            Top-level Result-Code from the CCA. Zero when skipped or
            when a transport error occurred.
        assertionsPassed:
          type: boolean
          description: True when all assertions on the step evaluated to true.
        assertions:
          type: array
          items:
            type: object
            required: [expression, passed]
            properties:
              expression:
                type: string
              passed:
                type: boolean
              message:
                type: string
                description: Non-empty when passed is false.
...
    ExecutionSummary:
      type: object
      required: [id, scenarioId, scenarioName, mode, state, startedAt]
      properties:
        id:
          type: string
        scenarioId:
          type: string
          description: The scenario this execution was started from.
        scenarioName:
          type: string
          description: Name at execution time (denormalised).
        mode:
          $ref: '#/components/schemas/ExecutionMode'
        peerId:
          type: string
        peerName:
          type: string
        subscriberId:
          type: string
        subscriberMsisdn:
          type: string
        batchId:
          type: string
          description: |
            Present on executions spawned from a batched Run.
        state:
          $ref: '#/components/schemas/ExecutionState'
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          description: Present only for terminal states.
    ExecutionPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionSummary'
        page:
          $ref: '#/components/schemas/PageMeta'
...